Add 16-bit Bayer pixel format and codec infrastructure
- PIXEL_FORMAT_RAW_RGGB_16 and GBRG_16 formats
- Updated wavelet, companding, and log curve for 16-bit range
- Prescale values {2,3,3} for 16-bit (vs {0,2,2} for 14-bit)
- Component clamping in decoder for Q6-Q8 overflow prevention
- PutBuffer overflow handling (assert → error return)

+++ source/lib/vc5_common/image.h
@@ -21,8 +21,8 @@
 #ifndef IMAGE_H
 #define IMAGE_H
 
-//! Data type for the values in a component array
-typedef uint16_t COMPONENT_VALUE;
+//! Data type for the values in a component array (widened for 16-bit pipeline)
+typedef int32_t COMPONENT_VALUE;
 
 /*!
 	@brief Data structure for an image input to the unpacking process
